requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.25.8
anyio==4.3.0
arxiv==2.1.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
ctransformers==0.2.27
cymem==2.0.8
dataclasses-json==0.6.4
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.16
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.0
faiss-cpu==1.8.0
feedparser==6.0.10
filelock==3.13.4
fireworks-ai==0.13.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.47.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.16
langchain-anthropic==0.1.11
langchain-community==0.0.33
langchain-core==0.1.43
langchain-openai
langchain-fireworks==0.1.2
langchain-google-genai==1.0.2
langchain-google-vertexai==1.0.1
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langcodes==3.4.0
langsmith==0.1.46
language_data==1.2.0
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
openai==1.17.1
orjson==3.10.0
packaging==23.2
pandas==2.2.2
pillow==10.3.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
python-box==7.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.6.1
sgmllib3k==1.0.0
shapely==2.0.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
streamlit==1.33.0
sympy==1.12
tavily-python==0.3.3
tenacity==8.2.3
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
transformers==4.39.3
typer==0.9.4
types-protobuf==4.25.0.20240410
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
wasabi==1.1.2
watchdog==4.0.0
weasel==0.3.4
wikipedia==1.4.0
xmltodict==0.13.0
yarl==1.9.4
